start.S: define sizes for FIP and BL1
Define and use sizes for the space reserved for the embedded FIP and
BL1 images.  (For BL1, the size is the sum of its RO and RW regions.)
This aligns better with the symbols representing the same values used
in ARM Trusted Firmware and U-Boot, and is moving toward using a
common set of definitions for these values.

Add a comment about an assumption that these offsets and sizes are
page aligned.  With that, we can remove some alignment directives.

Reserve all of the space claimed by the size stated for the BL1
image rather than just having the code that follows be aligned on
the next 256 byte boundary.

Signed-off-by: Alex Elder <elder@linaro.org>
Reviewed-by: Daniel Thompson <daniel.thompson@linaro.org>