l-loader: move FIP later in image
Move the FIP so it's at the end of the l-loader image rather than
in the middle.  This allows the offset used for BL1 (and BL2 and
BL31 that follow it) to be lower in memory.  The FIP is defined in
its own section to allow us to more directly specify the FIP's new
base memory address.

Signed-off-by: Alex Elder <elder@linaro.org>