Makefile: preprocess the loader script
Rename "l-loader.lds" to be "l-loader.ld.in" and use the renamed file
as input to the C Preprocessor to generate the actual loader script.
This will be used to allow symbolic values to define offsets, and
eventually to have those symbolic values be incorporated via #include.
(Having the preprocessor get this file as input from stdin eliminates
a warning about not using the linker input file.)

Signed-off-by: Alex Elder <elder@linaro.org>