Makefile: don't define TEXT_BASE in Makefile
Rather than passing it as an argument at build time, define the
TEXT_BASE address as LLOADER_TEXT_BASE (symbolically) inside the
linker script.  Add a comment to make explicit an alignment
requirement placed on this value. 

Similarly define it inside "start.S", so we can remove its use from
the Makefile.  It's temporarily defined in two places as a result;
this issue will be resolved in an upcoming patch (in which this
definition moves to a common include file).

Signed-off-by: Alex Elder <elder@linaro.org>