start.S: rename size and position symbols
Rename the symbols that define the offset within "l-loader" and the
size of regions of memory used by ARM Trusted Firmware.  This is to
prepare for the next patch, which takes these definitions from a
separate file (defined in the "arm-trusted-firmware" tree).  Most of
the symbols were previously named "CONFIG_*", but there is no need
to follow that naming convention (which was inherited from some
HiSilicon code this was derived from).

Signed-off-by: Alex Elder <elder@linaro.org>