start.S: safer wait loop
At the very beginning of reset(), there is a loop waiting until "the
system is ready."  The loop consists of repeatedly writing a value
to the REG_SC_GEN1 register (offset 0x84 the SCTL space at 0xf8000000)
until the value read back from the address matches what was written.
The power-on/reset value of that register is 0.

Currently, the value written is 1 initially, and it increments each
time through the loop.  The only problem with this is that if the
write never succeeds, or the "system" never becomes "ready", the
counter value could eventually wrap around to 0, and as a result,
the wait loop will end when it should not.

Fix this by simply writing a single non-zero value (1) rather than
incrementing a counter.

Signed-off-by: Alex Elder <elder@linaro.org>